[notion] add support for admiral panels
Former-commit-id: e28ce9f6845785bb7e9eca474c17646bc4c58bba [formerly 8aabd4a97051a8e5b9ef325e1d9b04e14d4da45c] [formerly b8ae22c91742ccc966f3b9ed6dd21b1582516e78 [formerly a1a070b649222be36f8dde4989430139ea000e82]]
Former-commit-id: 474ff234ef16087a6c6f3c40452c5a406162fd09 [formerly 1037985ea11c1d0356b54ba3391b5edc7273a963]
Former-commit-id: b22dbd77f1b99fb19b044c5d58105a2291592b5b

+++ .notion/cfg_dzen.lua
@@ -1,4 +1,5 @@
 home_=os.getenv("HOME")
+notion_dzen_=false
 
 dmain = {
     ws   = nil,
@@ -8,26 +9,31 @@ dmain = {
 }
 
 function dzen_update()
-    local template = ""
-    if dmain.date then
-        template = template.."^pa(2;)^bg(#000000)^ba()"..wrp(dmain.date).."^ba()^bg()"
+    if notion_dzen_ then
+        local template = ""
+        if dmain.date then
+            template = template.."^pa(2;)^bg(#000000)^ba()"..wrp(dmain.date).."^ba()^bg()"
+        end
+        if dmain.ws then
+            template = template..wrp(dmain.ws)
+        end
+        if dmain.kbd then
+            template = template..wrp(dmain.kbd)
+        end
+        if dmain.net then 
+            template = template..wrp("net: " .. dmain.net)
+        end
+        dzen_pipe:write(template..'\n')
     end
-    if dmain.ws then
-        template = template..wrp(dmain.ws)
-    end
-    if dmain.kbd then
-        template = template..wrp(dmain.kbd)
-    end
-    if dmain.net then 
-        template = template..wrp("net: " .. dmain.net)
-    end
-    dzen_pipe:write(template..'\n')
 end
 
 function dzen_delete()
     os.execute("pkill dzen2")
     os.execute("pkill rofi")
-    os.execute("pkill admiral")
+    if not notion_dzen_ then
+        os.execute("pkill admiral")
+        os.execute(home_.."/bin/scripts/panels")
+    end
 end
 
 local hook_deinit = notioncore.get_hook("ioncore_deinit_hook")
@@ -35,16 +41,18 @@ if hook_deinit then
     hook_deinit:add(dzen_delete)
 end
 
-if dzen_pipe == nil then
-    dzen_pipe = io.popen("dzen2 -dock -bg ".. 
-                        neg.dzen.bg_ ..
-                        " -h ".. neg.dzen.h_ ..
-                        " -tw 0 -x 0 -ta l -w ".. 
-                        neg.dzen.main_w_ .. 
-                        " -p -fn ".. 
-                        neg.dzen.panel_font_ .. 
-                        " ", "w"
-                        ) dzen_pipe:setvbuf("line")
+if notion_dzen_ then
+    if dzen_pipe == nil then
+        dzen_pipe = io.popen("dzen2 -dock -bg ".. 
+                            neg.dzen.bg_ ..
+                            " -h ".. neg.dzen.h_ ..
+                            " -tw 0 -x 0 -ta l -w ".. 
+                            neg.dzen.main_w_ .. 
+                            " -p -fn ".. 
+                            neg.dzen.panel_font_ .. 
+                            " ", "w"
+                            ) dzen_pipe:setvbuf("line")
+    end
 end
 
 dopath("dzen/ws")
